backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pandas>=2.2.0
httpx>=0.27.0
python-multipart>=0.0.6
pydantic>=2.6.0
python-dotenv>=1.0.0
